+++ src/constants/apiEndpoints.ts
@@ -1,0 +1,7 @@
+export class ApiEndpoints {
+    static readonly BASE_URL = "https://staging.debricked.com";
+
+    static getBaseUrl() {
+        return this.BASE_URL;
+    }
+}
