+++ src/constants/messages.ts
@@ -1,4 +1,5 @@
 export class Messages {
+    static readonly ACCESS_TOKEN = "Debricked access token";
     static readonly ENTER_ACCESS_TOKEN = "Enter your Debricked access token";
     static readonly ACCESS_TOKEN_RQD = "Access token is required";
     static readonly WS_NOT_FOUND = "No workspace folder found";
